week-4/load_data_to_gs_bq.requirements.txt
google-cloud-bigquery==3.6.0
google-cloud-storage==2.7.0
pandas==1.5.3
pyarrow==11.0.0
